Moves RenderTextureComponent to component package
