+++ src/pyrite/_rendering/render_texture.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
-from typing import cast, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from pygame import Rect, Surface
-
-from pyrite._component.component import BaseComponent
-from pyrite._types.protocols import HasTexture
 
 if TYPE_CHECKING:
     from pygame.typing import Point
@@ -54,19 +51,3 @@ class RenderTexture:
 
     def _resize_surface(self, size: Point) -> Surface:
         return Surface(size)
-
-
-class RenderTextureComponent(BaseComponent):
-    """
-    Special component for sprites and sprite-like objects. Automatically updates the
-    object's texture to the assigned rendertexture, and ensures it is redrawn to update.
-    """
-
-    def __init__(self, owner: HasTexture, render_texture: RenderTexture) -> None:
-        super().__init__(owner)
-        self.render_texture = render_texture
-
-    def update_texture(self):
-        owner = cast(HasTexture, self.owner)
-        owner.texture = self.render_texture.get_target_surface()
-        owner.is_dirty = True
