+++ src/pyrite/rendering.py
@@ -3,6 +3,7 @@ import pyrite._rendering.camera_renderer
 import pyrite._rendering.ortho_projection
 import pyrite._rendering.rect_bounds
 import pyrite._rendering.render_texture
+import pyrite._component.render_texture_component
 import pyrite._rendering.sprite_renderer
 import pyrite._rendering.viewport
 
@@ -11,6 +12,8 @@ CameraRenderer = pyrite._rendering.camera_renderer.CameraRendererProvider
 OrthoProjection = pyrite._rendering.ortho_projection.OrthoProjection
 RectBounds = pyrite._rendering.rect_bounds.RectBounds
 RenderTexture = pyrite._rendering.render_texture.RenderTexture
-RenderTextureComponent = pyrite._rendering.render_texture.RenderTextureComponent
+RenderTextureComponent = (
+    pyrite._component.render_texture_component.RenderTextureComponent
+)
 SpriteRenderer = pyrite._rendering.sprite_renderer.SpriteRendererProvider
 Viewport = pyrite._rendering.viewport.Viewport
